+++ src/main/java/DAO/RecipeDAO.java
@@ -21,7 +21,7 @@ public class RecipeDAO extends DBContext {
             while (rs.next()) {
                 int foodId = rs.getInt("FoodID");
                 int ingredientId = rs.getInt("IngredientID");
-                int requiredQuantity = rs.getInt("RequiredQuantity");
+                double requiredQuantity = rs.getDouble("RequiredQuantity"); // Thay getInt thành getDouble
 
                 list.add(new Recipe(foodId, ingredientId, requiredQuantity));
             }
@@ -41,7 +41,7 @@ public class RecipeDAO extends DBContext {
             ResultSet rs = ps.executeQuery();
             while (rs.next()) {
                 int ingredientId = rs.getInt("IngredientID");
-                int requiredQuantity = rs.getInt("RequiredQuantity");
+                double requiredQuantity = rs.getDouble("RequiredQuantity"); // Thay getInt thành getDouble
 
                 list.add(new Recipe(foodId, ingredientId, requiredQuantity));
             }
@@ -64,7 +64,7 @@ public class RecipeDAO extends DBContext {
             while (rs.next()) {
                 int foodId = rs.getInt("FoodID");
                 int ingredientId = rs.getInt("IngredientID");
-                int requiredQuantity = rs.getInt("RequiredQuantity");
+                double requiredQuantity = rs.getDouble("RequiredQuantity"); // Thay getInt thành getDouble
 
                 list.add(new Recipe(foodId, ingredientId, requiredQuantity));
             }
@@ -83,7 +83,7 @@ public class RecipeDAO extends DBContext {
             ps.setInt(2, ingredientId);
             ResultSet rs = ps.executeQuery();
             if (rs.next()) {
-                int requiredQuantity = rs.getInt("RequiredQuantity");
+                double requiredQuantity = rs.getDouble("RequiredQuantity"); // Thay getInt thành getDouble
                 return new Recipe(foodId, ingredientId, requiredQuantity);
             }
         } catch (SQLException e) {
@@ -99,7 +99,7 @@ public class RecipeDAO extends DBContext {
             PreparedStatement ps = conn.prepareStatement(query);
             ps.setInt(1, recipe.getFoodID());
             ps.setInt(2, recipe.getIngredientID());
-            ps.setInt(3, recipe.getRequiredQuantity());
+            ps.setDouble(3, recipe.getRequiredQuantity()); // Thay setInt thành setDouble
 
             return ps.executeUpdate() > 0;
         } catch (SQLException e) {
@@ -113,7 +113,7 @@ public class RecipeDAO extends DBContext {
         try {
             Connection conn = getConnection();
             PreparedStatement ps = conn.prepareStatement(query);
-            ps.setInt(1, recipe.getRequiredQuantity());
+            ps.setDouble(1, recipe.getRequiredQuantity()); // Thay setInt thành setDouble
             ps.setInt(2, recipe.getFoodID());
             ps.setInt(3, recipe.getIngredientID());
 
@@ -141,7 +141,6 @@ public class RecipeDAO extends DBContext {
         return false;
     }
 
-    // DAO hỗ trợ lấy tên Food và tên Ingredient
     public String getFoodNameById(int foodId) {
         String query = "SELECT FoodName FROM Foods WHERE FoodID = ?";
         try {
@@ -188,14 +187,14 @@ public class RecipeDAO extends DBContext {
             PreparedStatement ps = conn.prepareStatement(query);
             ResultSet rs = ps.executeQuery();
             while (rs.next()) {
-                Object[] row = new Object[7]; // Tăng kích thước array lên 7 để chứa thêm Image
+                Object[] row = new Object[7];
                 row[0] = rs.getInt("FoodID");
                 row[1] = rs.getInt("IngredientID");
-                row[2] = rs.getInt("RequiredQuantity");
+                row[2] = rs.getDouble("RequiredQuantity"); // Thay getInt thành getDouble
                 row[3] = rs.getString("FoodName");
                 row[4] = rs.getString("IngredientName");
                 row[5] = rs.getString("Unit");
-                row[6] = rs.getString("Image"); // Thêm trường Image
+                row[6] = rs.getString("Image");
 
                 list.add(row);
             }
@@ -219,14 +218,14 @@ public class RecipeDAO extends DBContext {
             ps.setString(1, "%" + foodName + "%");
             ResultSet rs = ps.executeQuery();
             while (rs.next()) {
-                Object[] row = new Object[7]; // Tăng kích thước array lên 7 để chứa thêm Image
+                Object[] row = new Object[7];
                 row[0] = rs.getInt("FoodID");
                 row[1] = rs.getInt("IngredientID");
-                row[2] = rs.getInt("RequiredQuantity");
+                row[2] = rs.getDouble("RequiredQuantity"); // Thay getInt thành getDouble
                 row[3] = rs.getString("FoodName");
                 row[4] = rs.getString("IngredientName");
                 row[5] = rs.getString("Unit");
-                row[6] = rs.getString("Image"); // Thêm trường Image
+                row[6] = rs.getString("Image");
 
                 list.add(row);
             }
@@ -236,7 +235,6 @@ public class RecipeDAO extends DBContext {
         return list;
     }
 
-    // Lấy danh sách tất cả Food
     public List<Object[]> getAllFoods() {
         List<Object[]> list = new ArrayList<>();
         String query = "SELECT FoodID, FoodName FROM Foods ORDER BY FoodName";
@@ -257,7 +255,6 @@ public class RecipeDAO extends DBContext {
         return list;
     }
 
-    // Lấy danh sách tất cả Ingredient
     public List<Object[]> getAllIngredients() {
         List<Object[]> list = new ArrayList<>();
         String query = "SELECT IngredientID, IngredientName, Unit FROM Ingredients ORDER BY IngredientName";
